requirements.txt
awesomeversion==25.5.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
Deprecated==1.2.18
idna==3.10
lxml==5.4.0
nvchecker==2.18
platformdirs==4.3.8
pycparser==2.22
pycurl==7.45.6
PyGithub==2.6.1
PyJWT==2.10.1
PyNaCl==1.5.0
python-frontmatter==1.1.0
pytoml==0.1.21
PyYAML==6.0.2
requests==2.32.4
structlog==25.4.0
toml==0.10.2
tornado==6.5.1
typing_extensions==4.14.0
urllib3==2.4.0
wrapt==1.17.2
